# backend/openapi_temp.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZiggyAI Trading Platform",
    "description": "Complete AI-powered trading platform with market analysis, news, signals, and execution",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/learning/data/summary": {
      "get": {
        "summary": "Get Learning Data Summary",
        "description": "Get summary of available learning data.",
        "operationId": "get_learning_data_summary_api_learning_data_summary_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 90,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/results/history": {
      "get": {
        "summary": "Get Learning History",
        "description": "Get history of learning results.",
        "operationId": "get_learning_history_api_learning_results_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/learning/evaluate/current": {
      "get": {
        "summary": "Evaluate Current Performance",
        "description": "Evaluate current rule performance.",
        "operationId": "evaluate_current_performance_api_learning_evaluate_current_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 180,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/learning/gates": {
      "get": {
        "summary": "Get Learning Gates",
        "description": "Get current validation gates configuration.",
        "operationId": "get_learning_gates_api_learning_gates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "put": {
        "summary": "Update Learning Gates",
        "description": "Update validation gates configuration.",
        "operationId": "update_learning_gates_api_learning_gates_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StrictGatesModel" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/calibration/build": {
      "post": {
        "summary": "Build Calibration Model",
        "description": "Build probability calibration model from recent data.",
        "operationId": "build_calibration_model_api_learning_calibration_build_post",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 30,
              "default": 90,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrictGatesModel": {
        "properties": {
          "min_trades": {
            "type": "integer",
            "title": "Min Trades",
            "default": 200
          },
          "min_sharpe_improvement_abs": {
            "type": "number",
            "title": "Min Sharpe Improvement Abs",
            "default": 0.2
          },
          "min_sharpe_improvement_rel": {
            "type": "number",
            "title": "Min Sharpe Improvement Rel",
            "default": 0.15
          },
          "max_brier_score_improvement": {
            "type": "number",
            "title": "Max Brier Score Improvement",
            "default": 0.02
          },
          "calibration_slope_range": {
            "prefixItems": [{ "type": "number" }, { "type": "number" }],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Calibration Slope Range",
            "default": [0.8, 1.2]
          },
          "max_drawdown_deterioration_rel": {
            "type": "number",
            "title": "Max Drawdown Deterioration Rel",
            "default": 0.1
          },
          "hit_rate_significance_p": {
            "type": "number",
            "title": "Hit Rate Significance P",
            "default": 0.05
          },
          "max_psi_threshold": {
            "type": "number",
            "title": "Max Psi Threshold",
            "default": 0.25
          },
          "max_daily_turnover_cap": {
            "type": "number",
            "title": "Max Daily Turnover Cap",
            "default": 50.0
          }
        },
        "type": "object",
        "title": "StrictGatesModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}